+++ packages/roosterjs-editor-api/lib/index.ts
@@ -40,3 +40,5 @@ export { default as toggleHeader } from './format/toggleHeader';
 export { default as defaultLinkMatchRules } from './linkMatch/defaultLinkMatchRules';
 export { default as matchLink } from './linkMatch/matchLink';
 export { default as RegExLinkMatchRule } from './linkMatch/RegExLinkMatchRule';
+export { default as buildClipBoardData } from './paste/buildClipBoardData';
+export { getPastedElementWithId } from './paste/utils/processImages';
