+++ packages/roosterjs-editor-types/lib/editor/ClipboardData.ts
@@ -20,6 +20,7 @@ interface ClipBoardData {
     imageData: ImageData;
     textData: string;
     htmlData: string;
+    htmlFragment: DocumentFragment;
 }
 
 export default ClipBoardData;
