+++ packages/roosterjs-editor-types/lib/index.ts
@@ -23,3 +23,9 @@ export { default as PluginEventType } from './editor/PluginEventType';
 export { default as Rect } from './editor/Rect';
 export { default as TextSection } from './editor/TextSection';
 export { default as TraversingScoper } from './editor/TraversingScoper';
+export {
+    default as ClipBoardData,
+    ImageData,
+    LocalFileImageData,
+    Base64ImageData,
+} from './editor/ClipboardData';
